[fix] 메인페이지 질문 코드 수정
