+++ src/components/common/MiniToggle.tsx
@@ -1,10 +1,11 @@
 import styled from 'styled-components'
-import { colors } from '../../styles/colors'
+import { colors } from '@/styles/colors'
 
 interface MiniToggleProps {
   isActive: boolean
   setIsActive: React.Dispatch<React.SetStateAction<boolean>>
 }
+
 const MiniToggle = ({ isActive, setIsActive }: MiniToggleProps) => {
   return (
     <ToggleSwitch>
@@ -17,15 +18,14 @@ const MiniToggle = ({ isActive, setIsActive }: MiniToggleProps) => {
 export default MiniToggle
 
 const ToggleSwitch = styled.label`
-  position: relative;
   display: inline-block;
+  position: relative;
   width: 36px;
   height: 19px;
 `
 
 const ToggleSlider = styled.span`
   position: absolute;
-  cursor: pointer;
   top: 0;
   left: 0;
   right: 0;
@@ -34,6 +34,7 @@ const ToggleSlider = styled.span`
   -webkit-transition: 0.4s;
   transition: 0.4s;
   border-radius: 100px;
+  cursor: pointer;
 
   &:before {
     position: absolute;
