+++ src/components/folder/FolderList.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react'
 import EachFolder from '@/components/folder/EachFolder'
 import { colors } from '@/styles/colors'
 import { UnFixedButton } from '@/components/common/Button'
-import { directoryProps } from '../main/types'
+import { directoryProps } from '../question/types'
 import { useRecoilValue } from 'recoil'
 import { UserInfoStateProps, userInfoState } from '@/context/Atoms'
 import { getDirectoriesApi, makeDirectoryApi } from '@/apis/DirectoryApi'
