+++ src/components/folder/types.tsx
@@ -1,0 +1,7 @@
+// 디렉토리 Props
+export interface directoryProps {
+  categoryId: number
+  categoryName: string
+  answerIds: number[]
+  categoryImage: string
+}
