+++ src/components/question/types.tsx
@@ -1,0 +1,11 @@
+// userData Props
+export interface userDataProps {
+  nickname: string
+  memberId: number
+}
+
+// 모달창 Props
+export interface ModalProps {
+  content: string
+  clickModal: () => void
+}
