+++ src/components/toast/toastStyle.ts
@@ -1,6 +1,6 @@
 import { ToastContainer } from 'react-toastify'
 import styled from 'styled-components'
-import { colors } from '../../styles/colors'
+import { colors } from '@/styles/colors'
 
 export const StyledToastContainer = styled(ToastContainer)`
   .Toastify__toast {
@@ -8,13 +8,12 @@ export const StyledToastContainer = styled(ToastContainer)`
     justify-content: center;
     align-items: center;
     border-radius: 12px;
+    margin: 0px 20px 30px 20px;
     text-align: center;
     background: rgba(29, 29, 29, 0.8);
     color: ${colors.white};
     font-family: Pretendard;
     font-size: 16px;
-    margin: 0px 20px 30px 20px;
-    font-style: normal;
     font-weight: 600;
     line-height: 24px;
     letter-spacing: -0.32px;
